hw05 spring library v1.2

+++ src/main/java/ru/otus/library/domain/Author.java
@@ -1,23 +1,23 @@
 package ru.otus.library.domain;
 
 public class Author {
-    private int id=0;
+    private long id=0;
     private String fio;
 
     public Author(String fio) {
         this.fio = fio;
     }
 
-    public Author(int id, String fio) {
+    public Author(long id, String fio) {
         this.id = id;
         this.fio = fio;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
